refactor: 패키지 구조 객체에 집중할 수 있는 방향으로 수정
