+++ src/main/java/roomescape/reservation/domain/ReservationDateTime.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation;
+package roomescape.reservation.domain;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
