+++ src/main/java/roomescape/reservation/domain/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation;
+package roomescape.reservation.domain;
 
 import java.util.List;
