+++ src/main/java/roomescape/reservation/dto/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.reservation.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
