+++ src/main/java/roomescape/reservation/infrastructure/MemoryReservationRepository.java
@@ -1,12 +1,12 @@
-package roomescape.infrastructure.reservation;
+package roomescape.reservation.infrastructure;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.ReservationRepository;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.domain.ReservationRepository;
 
 @Repository
 public class MemoryReservationRepository implements ReservationRepository {
