+++ src/main/java/roomescape/time/dto/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.time.dto;
 
 import java.time.LocalTime;
